fix ghostbuster
gems are installed using bundler.
If we want to use an extension installed via bundler, we must run
bundler.
To do so we need the Gemfile which was used to build the container.

Besides this the execution of ghostbuster must be done using bundle exec
puppet.lint instead of puppet-lint.